requirements.txt
openai
pandas
numpy
python-dotenv
scikit-learn
rank_bm25
